fix: Fix issues suggestions when lineNumber is returned CY-4494
On the suggestion informationt that comes from the tool, there can be a line number,
and that can be a problem, especially if the line number is different
than that issue number. Example:

{"filename":"readme.md","message":"Expected: 1; Actual: 0; Below","patternId":"MD022","line":3,"suggestion":"\n### Among Codacy’s features"}
Fix Information:  { lineNumber: 4, insertText: '\n' }

In the case above, we insert the '\n' on the begining of the line 4, but it should be inserted
at the end of line 3. Since we only have one "line", the line of the issue and suggestion must match.
For cases where we cannotassure that, we should not give any suggestion instead