Remove generateDocs from scripts in main project.
Its not really workable and is broken.
The way to generate the docs is to go into its directory,
run compile and node/dist/index.js as is in the docker file.
Reason being that in a different directory it doesnt find
the docs directory properly since is relative when searching
for it.